backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
reportlab==4.2.2
openai==1.51.2
python-multipart==0.0.9
selenium==4.25.0
webdriver-manager==4.0.2
pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.0
httpx==0.27.0